Clase 24 add methods Dentist.search, Dentist.update, Dentist.delete, Patient.register , Address.register

+++ Clase_25A_SpringBoot_API_Exercise/src/main/java/com/example/Clase_25A_SpringBoot_API_Exercise/dao/Idao.java
@@ -6,9 +6,9 @@ public interface Idao<T> {
 
     public T register(T t);
 
-    public void delete(long id);
+    public void delete(int id);
 
-    public T search(long id);
+    public T search(int id);
 
     public List<T> searchAll();
